requirements.txt
pandas
numpy
yfinance
matplotlib
seaborn
scipy
jupyterlab
tqdm
statsmodels
pyyaml